+++ community/src/main/java/org/community/annotation/CurrentPost.java
@@ -1,4 +1,10 @@
 package org.community.annotation;
 
+import java.lang.annotation.*;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
 public @interface CurrentPost {
+
 }
