+++ community/src/main/java/org/community/service/post/PostService.java
@@ -1,5 +1,6 @@
 package org.community.service.post;
 
+import jakarta.annotation.Nullable;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -28,104 +29,16 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-@Transactional
-@Slf4j
-public class PostService {
-    private final PostRepository postRepository;
-    private final UserRepository userRepository;
-    private final CommentService commentService;
-    private final LikeService likeService;
-    private final JwtUtil jwtUtil;
-
-    public ResponseEntity<ApiResponse> getPosts(int page, int size) {
-        Pageable pageable = PageRequest.of(page, size, Sort.by("regDate").descending());
-        Page<PostEntity> pages = postRepository.findAll(pageable);
-
-        List<PostResponse> postList = pages.getContent().stream()
-                .map(PostResponse::new)
-                .collect(Collectors.toList());
-
-        return ApiResponse.response(UserResponseMessage.POST_FETCH_SUCCESS, postList);
-    }
-
-
-    public ResponseEntity<ApiResponse> createPosts(HttpServletRequest request, PostCreateRequest postCreateRequest, String imagePath) {
-        Long userId = jwtUtil.getUserIdFromJwt(request.getHeader("Authorization"));
-        UserEntity user = userRepository.findById(userId).orElseThrow(() ->
-                new CustomException(UserResponseMessage.USER_NOT_FOUND)
-        );
-        PostEntity newPost = postCreateRequest.toEntity(user);
-        newPost.setPostImagePath(imagePath);
-        postRepository.save(newPost);
-        return ApiResponse.response(UserResponseMessage.POST_CREATED);
-    }
-
-    public ResponseEntity<ApiResponse> getPostDetail(HttpServletRequest request, Long postId) {
-        PostEntity post = postRepository.findById(postId).orElseThrow(() -> new CustomException(UserResponseMessage.POST_NOT_FOUND));
-        Long userId = jwtUtil.getUserIdFromJwt(request.getHeader("Authorization"));
-        PostDetailResponse responseBody = new PostDetailResponse(post, userId);
-
-        List<Comment> commentList = commentService.getCommentsByPost(userId, post);
-        responseBody.setCommentList(commentList);
-
-        Boolean isLiked = likeService.getIsLiked(request, post);
-        responseBody.setIsLiked(isLiked);
-        post.setViews(post.getViews() + 1);
-        return ApiResponse.response(UserResponseMessage.POST_FETCH_SUCCESS, responseBody);
-    }
-
-    public ResponseEntity<ApiResponse> updatePost(Long postId, PostUpdateRequest postUpdateRequest, String imagePath) {
-
-        PostEntity post = postRepository.findById(postId).orElseThrow(() ->
-                new CustomException(UserResponseMessage.POST_NOT_FOUND)
-        );
-
-        // 기존 이미지 파일 삭제
-        String fullOldImagePath = System.getProperty("user.dir") + "/community" + post.getPostImagePath();
-        File oldFile = new File(fullOldImagePath);
-
-        if (oldFile.exists()) {
-            boolean deleted = oldFile.delete();
-            if (!deleted) {
-                log.warn("기존 이미지 파일 삭제 실패: {}", fullOldImagePath);
-            } else {
-                log.info("기존 이미지 파일 삭제 성공: {}", fullOldImagePath);
-            }
-        }
-
-        post.setTitle(postUpdateRequest.getTitle());
-        post.setContents(postUpdateRequest.getContents());
-        post.setPostImagePath(imagePath);
-
-        return ApiResponse.response(UserResponseMessage.POST_UPDATED);
-    }
-
-    /// ////////
-
-    public ResponseEntity<ApiResponse> deletePost(Long postId) {
-        PostEntity post = postRepository.findById(postId).orElseThrow(() -> new CustomException(UserResponseMessage.POST_NOT_FOUND));
-
-        // 기존 이미지 파일 삭제
-        String fullOldImagePath = System.getProperty("user.dir") + "/community" + post.getPostImagePath();
-        File oldFile = new File(fullOldImagePath);
-
-        if (oldFile.exists()) {
-            boolean deleted = oldFile.delete();
-            if (!deleted) {
-                log.warn("기존 이미지 파일 삭제 실패: {}", fullOldImagePath);
-            } else {
-                log.info("기존 이미지 파일 삭제 성공: {}", fullOldImagePath);
-            }
-        }
-
-        postRepository.delete(post);
-        return ApiResponse.response(UserResponseMessage.POST_DELETED);
-    }
+public interface PostService {
+    ResponseEntity<ApiResponse> getPosts(int page, int size);
+    ResponseEntity<ApiResponse> createPosts(UserEntity user, PostCreateRequest postCreateRequest, @Nullable MultipartFile postImage);
+    ResponseEntity<ApiResponse> getPostDetail(HttpServletRequest request, Long postId);
+    ResponseEntity<ApiResponse> updatePost(Long postId, PostUpdateRequest postUpdateRequest, String imagePath);
+    ResponseEntity<ApiResponse> deletePost(Long postId);
 }
